+++ clients/admin-ui/src/pages/settings/custom-fields/[id].tsx
@@ -1,30 +1,19 @@
-import { AntSkeleton as Skeleton } from "fidesui";
 import { useRouter } from "next/router";
 
 import Layout from "~/features/common/Layout";
 import { CUSTOM_FIELDS_ROUTE } from "~/features/common/nav/routes";
 import PageHeader from "~/features/common/PageHeader";
-import CustomFieldFormV2 from "~/features/custom-fields/CustomFieldForm";
+import CustomFieldForm from "~/features/custom-fields/CustomFieldForm";
 import { useGetCustomFieldDefinitionByIdQuery } from "~/features/plus/plus.slice";
-
-const SkeletonCustomFieldForm = () => {
-  return (
-    <Skeleton active>
-      <Skeleton.Input />
-      <Skeleton.Input />
-      <Skeleton.Input />
-      <Skeleton.Input />
-      <Skeleton.Button />
-    </Skeleton>
-  );
-};
 
 const CustomFieldDetailPage = () => {
   const router = useRouter();
   const { id } = router.query;
 
   const { data: customField, isLoading } = useGetCustomFieldDefinitionByIdQuery(
-    { id: id as string },
+    {
+      id: id as string,
+    },
   );
 
   return (
@@ -36,11 +25,7 @@ const CustomFieldDetailPage = () => {
           { title: customField?.name ?? id },
         ]}
       />
-      {isLoading ? (
-        <SkeletonCustomFieldForm />
-      ) : (
-        <CustomFieldFormV2 initialField={customField} />
-      )}
+      <CustomFieldForm initialField={customField} isLoading={isLoading} />
     </Layout>
   );
 };