+++ CHANGELOG.md
@@ -38,7 +38,12 @@ Changes can also be flagged with a GitHub label for tracking purposes. The URL o
 - Trigger fidesplus CI actions on pushes to `main` in Fides [#6592](https://github.com/ethyca/fides/pull/6592)
 - Improved readability and maintainability around DSR packages and improved test coverage [#6540](https://github.com/ethyca/fides/pull/6540)
 
-## [2.70.5](https://github.com/ethyca/fides/compare/2.70.3...2.70.5)
+## [2.70.6](https://github.com/ethyca/fides/compare/2.70.5...2.70.6)
+
+### Changed
+- Update default GPC description text [#6661](https://github.com/ethyca/fides/pull/6661)
+
+## [2.70.5](https://github.com/ethyca/fides/compare/2.70.4...2.70.5)
 
 ### Fixed
 - Encode GPC subsection in GPP string whenever supported by the section (e.g. usnat, usca) [#6633](https://github.com/ethyca/fides/pull/6633)